requirements.txt
# ========================================
# FILE: requirements.txt
# ========================================
Flask==3.1.2
Flask-CORS==6.0.1
pyserial==3.5
python-dotenv==1.2.0
requests==2.32.5